Corrected typo in Internal DNS Service page
Example demonstrates using `bar.stackb` but the actual text described using `foo.stackb`.

The casing changes too (e.g. `stackB`) but if my understanding is correct then DNS is case insensitive so it should not matter. Let me know if I am wrong (or if you would just like me to) and I can correct that in the example too.